adding datat

+++ src/main/java/com/codewithhrishi/entity/Order.java
@@ -35,7 +35,7 @@ public class Order {
 	@OneToMany
 	private List<OrderItem>items;
 	private int totalItem;
-	private int totalPrice;
+	private long totalPrice;
